separate out reactions
