+++ musica/mechanism_configuration.py
@@ -6,1346 +6,52 @@
 import os
 import json
 import yaml
-from typing import Optional, Any, Dict, List, Union, Tuple
+from typing import Optional, Any, Dict, List
 from musica import (
     # _ReactionType,
     # _Species,
     # _Phase,
-    _ReactionComponent,
-    _Arrhenius,
-    _CondensedPhaseArrhenius,
-    _Troe,
-    _Branched,
-    _Tunneling,
-    _Surface,
-    _Photolysis,
-    _CondensedPhasePhotolysis,
-    _Emission,
-    _FirstOrderLoss,
-    _AqueousEquilibrium,
-    _WetDeposition,
-    _HenrysLaw,
-    _SimpolPhaseTransfer,
-    _UserDefined,
+    # _ReactionComponent,
+    # _Arrhenius,
+    # _CondensedPhaseArrhenius,
+    # _Troe,
+    # _Branched,
+    # _Tunneling,
+    # _Surface,
+    # _Photolysis,
+    # _CondensedPhasePhotolysis,
+    # _Emission,
+    # _FirstOrderLoss,
+    # _AqueousEquilibrium,
+    # _WetDeposition,
+    # _HenrysLaw,
+    # _SimpolPhaseTransfer,
+    # _UserDefined,
     # _Reactions,
     # _ReactionsIterator,
     _Mechanism,
     _Version,
     _Parser,
 )
-from musica.mechanism_configuration_species import Species, _Species
+from musica.mechanism_configuration_species import Species
 from musica.mechanism_configuration_phase import Phase
-from musica.mechanism_configuration_reactions import ReactionType, Reactions, ReactionComponentSerializer
-from musica.mechanism_configuration_utils import add_other_properties, remove_empty_keys
-
-BOLTZMANN_CONSTANT_J_K = 1.380649e-23  # J K-1
-
-
-class Arrhenius(_Arrhenius):
-    """
-    A class representing an Arrhenius rate constant.
-
-    k = A * exp( C / T ) * ( T / D )^B * exp( 1 - E * P )
-
-    where:
-        k = rate constant
-        A = pre-exponential factor [(mol m-3)^(n-1)s-1]
-        B = temperature exponent [unitless]
-        C = exponential term [K-1]
-        D = reference temperature [K]
-        E = pressure scaling term [Pa-1]
-        T = temperature [K]
-        P = pressure [Pa]
-        n = number of reactants
-
-    Attributes:
-        name (str): The name of the Arrhenius rate constant.
-        A (float): Pre-exponential factor [(mol m-3)^(n-1)s-1] where n is the number of reactants.
-        B (float): Temperature exponent [unitless].
-        C (float): Exponential term [K-1].
-        D (float): Reference Temperature [K].
-        E (float): Pressure scaling term [Pa-1].
-        reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-        products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-        gas_phase (Phase): The gas phase in which the reaction occurs.
-        other_properties (Dict[str, Any]): A dictionary of other properties of the Arrhenius rate constant.
-    """
-
-    def __init__(
-        self,
-        name: Optional[str] = None,
-        A: Optional[float] = None,
-        B: Optional[float] = None,
-        C: Optional[float] = None,
-        Ea: Optional[float] = None,
-        D: Optional[float] = None,
-        E: Optional[float] = None,
-        reactants: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        products: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        gas_phase: Optional[Phase] = None,
-        other_properties: Optional[Dict[str, Any]] = None,
-    ):
-        """
-        Initializes the Arrhenius object with the given parameters.
-
-        Args:
-            name (str): The name of the Arrhenius rate constant.
-            A (float): Pre-exponential factor [(mol m-3)^(n-1)s-1] where n is the number of reactants.
-            B (float): Temperature exponent [unitless].
-            C (float): Exponential term [K-1].
-            Ea (float): Activation energy [J molecule-1].
-            D (float): Reference Temperature [K].
-            E (float): Pressure scaling term [Pa-1].
-            reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-            products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-            gas_phase (Phase): The gas phase in which the reaction occurs.
-            other_properties (Dict[str, Any]): A dictionary of other properties of the Arrhenius rate constant.
-        """
-        super().__init__()
-        self.name = name if name is not None else self.name
-        self.A = A if A is not None else self.A
-        self.B = B if B is not None else self.B
-        if C is not None and Ea is not None:
-            raise ValueError("Cannot specify both C and Ea.")
-        self.C = -Ea / BOLTZMANN_CONSTANT_J_K if Ea is not None else C if C is not None else self.C
-        self.D = D if D is not None else self.D
-        self.E = E if E is not None else self.E
-        self.reactants = (
-            [
-                (
-                    _ReactionComponent(r.name)
-                    if isinstance(r, Species)
-                    else _ReactionComponent(r[1].name, r[0])
-                )
-                for r in reactants
-            ]
-            if reactants is not None
-            else self.reactants
-        )
-        self.products = (
-            [
-                (
-                    _ReactionComponent(p.name)
-                    if isinstance(p, Species)
-                    else _ReactionComponent(p[1].name, p[0])
-                )
-                for p in products
-            ]
-            if products is not None
-            else self.products
-        )
-        self.gas_phase = gas_phase.name if gas_phase is not None else self.gas_phase
-        self.other_properties = other_properties if other_properties is not None else self.other_properties
-
-    @staticmethod
-    def serialize(cls):
-        serialize_dict = {
-            "type": "ARRHENIUS",
-            "name": cls.name,
-            "A": cls.A,
-            "B": cls.B,
-            "C": cls.C,
-            "D": cls.D,
-            "E": cls.E,
-            "reactants": ReactionComponentSerializer.serialize_list_reaction_components(cls.reactants),
-            "products": ReactionComponentSerializer.serialize_list_reaction_components(cls.products),
-            "gas phase": cls.gas_phase,
-        }
-        add_other_properties(serialize_dict, cls.other_properties)
-        return remove_empty_keys(serialize_dict)
-
-
-class CondensedPhaseArrhenius(_CondensedPhaseArrhenius):
-    """
-    A class representing a condensed phase Arrhenius rate constant.
-
-    Attributes:
-        name (str): The name of the condensed phase Arrhenius rate constant.
-        A (float): Pre-exponential factor [(mol m-3)^(n-1)s-1].
-        B (float): Temperature exponent [unitless].
-        C (float): Exponential term [K-1].
-        Ea (float): Activation energy [J molecule-1].
-        D (float): Reference Temperature [K].
-        E (float): Pressure scaling term [Pa-1].
-        reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-        products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-        aerosol_phase (Phase): The aerosol phase in which the reaction occurs.
-        aerosol_phase_water (Species): The water species in the aerosol phase.
-        other_properties (Dict[str, Any]): A dictionary of other properties of the condensed phase Arrhenius rate constant.
-    """
-
-    def __init__(
-        self,
-        name: Optional[str] = None,
-        A: Optional[float] = None,
-        B: Optional[float] = None,
-        C: Optional[float] = None,
-        Ea: Optional[float] = None,
-        D: Optional[float] = None,
-        E: Optional[float] = None,
-        reactants: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        products: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        aerosol_phase: Optional[Phase] = None,
-        aerosol_phase_water: Optional[Species] = None,
-        other_properties: Optional[Dict[str, Any]] = None,
-    ):
-        """
-        Initializes the CondensedPhaseArrhenius object with the given parameters.
-
-        Args:
-            name (str): The name of the condensed phase Arrhenius rate constant.
-            A (float): Pre-exponential factor [(mol m-3)^(n-1)s-1].
-            B (float): Temperature exponent [unitless].
-            C (float): Exponential term [K-1].
-            Ea (float): Activation energy [J molecule-1].
-            D (float): Reference Temperature [K].
-            E (float): Pressure scaling term [Pa-1].
-            reactants (List[Union[Species, Tuple[float, Species]]]]: A list of reactants involved in the reaction.
-            products (List[Union[Species, Tuple[float, Species]]]]: A list of products formed in the reaction.
-            aerosol_phase (Phase): The aerosol phase in which the reaction occurs.
-            aerosol_phase_water (Species): The water species in the aerosol phase.
-            other_properties (Dict[str, Any]): A dictionary of other properties of the condensed phase Arrhenius rate constant.
-        """
-        super().__init__()
-        self.name = name if name is not None else self.name
-        self.A = A if A is not None else self.A
-        self.B = B if B is not None else self.B
-        if C is not None and Ea is not None:
-            raise ValueError("Cannot specify both C and Ea.")
-        self.C = -Ea / BOLTZMANN_CONSTANT_J_K if Ea is not None else C if C is not None else self.C
-        self.D = D if D is not None else self.D
-        self.E = E if E is not None else self.E
-        self.reactants = (
-            [
-                (
-                    _ReactionComponent(r.name)
-                    if isinstance(r, Species)
-                    else _ReactionComponent(r[1].name, r[0])
-                )
-                for r in reactants
-            ]
-            if reactants is not None
-            else self.reactants
-        )
-        self.products = (
-            [
-                (
-                    _ReactionComponent(p.name)
-                    if isinstance(p, Species)
-                    else _ReactionComponent(p[1].name, p[0])
-                )
-                for p in products
-            ]
-            if products is not None
-            else self.products
-        )
-        self.aerosol_phase = aerosol_phase.name if aerosol_phase is not None else self.aerosol_phase
-        self.aerosol_phase_water = (
-            aerosol_phase_water.name if aerosol_phase_water is not None else self.aerosol_phase_water
-        )
-        self.other_properties = other_properties if other_properties is not None else self.other_properties
-
-    @staticmethod
-    def serialize(cls):
-        serialize_dict = {
-            "type": "CONDENSED_PHASE_ARRHENIUS",
-            "name": cls.name,
-            "A": cls.A,
-            "B": cls.B,
-            "C": cls.C,
-            "D": cls.D,
-            "E": cls.E,
-            "reactants": ReactionComponentSerializer.serialize_list_reaction_components(cls.reactants),
-            "products":  ReactionComponentSerializer.serialize_list_reaction_components(cls.products),
-            "aerosol phase": cls.aerosol_phase,
-            "aerosol-phase water": cls.aerosol_phase_water,
-        }
-        add_other_properties(serialize_dict, cls.other_properties)
-        return remove_empty_keys(serialize_dict)
-
-
-class Troe(_Troe):
-    """
-    A class representing a Troe rate constant.
-
-    Attributes:
-        name (str): The name of the Troe rate constant.
-        k0_A (float): Pre-exponential factor for the low-pressure limit [(mol m-3)^(n-1)s-1].
-        k0_B (float): Temperature exponent for the low-pressure limit [unitless].
-        k0_C (float): Exponential term for the low-pressure limit [K-1].
-        kinf_A (float): Pre-exponential factor for the high-pressure limit [(mol m-3)^(n-1)s-1].
-        kinf_B (float): Temperature exponent for the high-pressure limit [unitless].
-        kinf_C (float): Exponential term for the high-pressure limit [K-1].
-        Fc (float): Troe parameter [unitless].
-        N (float): Troe parameter [unitless].
-        reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-        products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-        gas_phase (Phase): The gas phase in which the reaction occurs.
-        other_properties (Dict[str, Any]): A dictionary of other properties of the Troe rate constant.
-    """
-
-    def __init__(
-        self,
-        name: Optional[str] = None,
-        k0_A: Optional[float] = None,
-        k0_B: Optional[float] = None,
-        k0_C: Optional[float] = None,
-        kinf_A: Optional[float] = None,
-        kinf_B: Optional[float] = None,
-        kinf_C: Optional[float] = None,
-        Fc: Optional[float] = None,
-        N: Optional[float] = None,
-        reactants: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        products: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        gas_phase: Optional[Phase] = None,
-        other_properties: Optional[Dict[str, Any]] = None,
-    ):
-        """
-        Initializes the Troe object with the given parameters.
-
-        k0 = k0_A * exp( k0_C / T ) * ( T / 300.0 )^k0_B
-        kinf = kinf_A * exp( kinf_C / T ) * ( T / 300.0 )^kinf_B
-        k = k0[M] / ( 1 + k0[M] / kinf ) * Fc^(1 + 1/N*(log10(k0[M]/kinf))^2)^-1
-
-        where:
-            k = rate constant
-            k0 = low-pressure limit rate constant
-            kinf = high-pressure limit rate constant
-            k0_A = pre-exponential factor for the low-pressure limit [(mol m-3)^(n-1)s-1]
-            k0_B = temperature exponent for the low-pressure limit [unitless]
-            k0_C = exponential term for the low-pressure limit [K-1]
-            kinf_A = pre-exponential factor for the high-pressure limit [(mol m-3)^(n-1)s-1]
-            kinf_B = temperature exponent for the high-pressure limit [unitless]
-            kinf_C = exponential term for the high-pressure limit [K-1]
-            Fc = Troe parameter [unitless]
-            N = Troe parameter [unitless]
-            T = temperature [K]
-            M = concentration of the third body [mol m-3]
-
-        Args:
-            name (str): The name of the Troe rate constant.
-            k0_A (float): Pre-exponential factor for the low-pressure limit [(mol m-3)^(n-1)s-1].
-            k0_B (float): Temperature exponent for the low-pressure limit [unitless].
-            k0_C (float): Exponential term for the low-pressure limit [K-1].
-            kinf_A (float): Pre-exponential factor for the high-pressure limit [(mol m-3)^(n-1)s-1].
-            kinf_B (float): Temperature exponent for the high-pressure limit [unitless].
-            kinf_C (float): Exponential term for the high-pressure limit [K-1].
-            Fc (float): Troe parameter [unitless].
-            N (float): Troe parameter [unitless].
-            reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-            products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-            gas_phase (Phase): The gas phase in which the reaction occurs.
-            other_properties (Dict[str, Any]): A dictionary of other properties of the Troe rate constant.
-        """
-        super().__init__()
-        self.name = name if name is not None else self.name
-        self.k0_A = k0_A if k0_A is not None else self.k0_A
-        self.k0_B = k0_B if k0_B is not None else self.k0_B
-        self.k0_C = k0_C if k0_C is not None else self.k0_C
-        self.kinf_A = kinf_A if kinf_A is not None else self.kinf_A
-        self.kinf_B = kinf_B if kinf_B is not None else self.kinf_B
-        self.kinf_C = kinf_C if kinf_C is not None else self.kinf_C
-        self.Fc = Fc if Fc is not None else self.Fc
-        self.N = N if N is not None else self.N
-        self.reactants = (
-            [
-                (
-                    _ReactionComponent(r.name)
-                    if isinstance(r, Species)
-                    else _ReactionComponent(r[1].name, r[0])
-                )
-                for r in reactants
-            ]
-            if reactants is not None
-            else self.reactants
-        )
-        self.products = (
-            [
-                (
-                    _ReactionComponent(p.name)
-                    if isinstance(p, Species)
-                    else _ReactionComponent(p[1].name, p[0])
-                )
-                for p in products
-            ]
-            if products is not None
-            else self.products
-        )
-        self.gas_phase = gas_phase.name if gas_phase is not None else self.gas_phase
-        self.other_properties = other_properties if other_properties is not None else self.other_properties
-
-    @staticmethod
-    def serialize(cls):
-        serialize_dict = {
-            "type": "TROE",
-            "name": cls.name,
-            "k0_A": cls.k0_A,
-            "k0_B": cls.k0_B,
-            "k0_C": cls.k0_C,
-            "kinf_A": cls.kinf_A,
-            "kinf_B": cls.kinf_B,
-            "kinf_C": cls.kinf_C,
-            "Fc": cls.Fc,
-            "N": cls.N,
-            "reactants": ReactionComponentSerializer.serialize_list_reaction_components(cls.reactants),
-            "products": ReactionComponentSerializer.serialize_list_reaction_components(cls.products),
-            "gas phase": cls.gas_phase,
-        }
-        add_other_properties(serialize_dict, cls.other_properties)
-        return remove_empty_keys(serialize_dict)
-
-
-class Branched(_Branched):
-    """
-    A class representing a branched reaction rate constant.
-
-    (TODO: get details from MusicBox)
-
-    Attributes:
-        name (str): The name of the branched reaction rate constant.
-        X (float): Pre-exponential branching factor [(mol m-3)^-(n-1)s-1].
-        Y (float): Exponential branching factor [K-1].
-        a0 (float): Z parameter [unitless].
-        n (float): A parameter [unitless].
-        reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-        nitrate_products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the nitrate branch.
-        alkoxy_products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the alkoxy branch.
-        gas_phase (Phase): The gas phase in which the reaction occurs.
-        other_properties (Dict[str, Any]): A dictionary of other properties of the branched reaction rate constant.
-    """
-
-    def __init__(
-        self,
-        name: Optional[str] = None,
-        X: Optional[float] = None,
-        Y: Optional[float] = None,
-        a0: Optional[float] = None,
-        n: Optional[float] = None,
-        reactants: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        nitrate_products: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        alkoxy_products: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        gas_phase: Optional[Phase] = None,
-        other_properties: Optional[Dict[str, Any]] = None,
-    ):
-        """
-        Initializes the Branched object with the given parameters.
-
-        Args:
-            name (str): The name of the branched reaction rate constant.
-            X (float): Pre-exponential branching factor [(mol m-3)^-(n-1)s-1].
-            Y (float): Exponential branching factor [K-1].
-            a0 (float): Z parameter [unitless].
-            n (float): A parameter [unitless].
-            reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-            nitrate_products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the nitrate branch.
-            alkoxy_products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the alkoxy branch.
-            gas_phase (Phase): The gas phase in which the reaction occurs.
-            other_properties (Dict[str, Any]): A dictionary of other properties of the branched reaction rate constant.
-        """
-        super().__init__()
-        self.name = name if name is not None else self.name
-        self.X = X if X is not None else self.X
-        self.Y = Y if Y is not None else self.Y
-        self.a0 = a0 if a0 is not None else self.a0
-        self.n = n if n is not None else self.n
-        self.reactants = (
-            [
-                (
-                    _ReactionComponent(r.name)
-                    if isinstance(r, Species)
-                    else _ReactionComponent(r[1].name, r[0])
-                )
-                for r in reactants
-            ]
-            if reactants is not None
-            else self.reactants
-        )
-        self.nitrate_products = (
-            [
-                (
-                    _ReactionComponent(p.name)
-                    if isinstance(p, Species)
-                    else _ReactionComponent(p[1].name, p[0])
-                )
-                for p in nitrate_products
-            ]
-            if nitrate_products is not None
-            else self.nitrate_products
-        )
-        self.alkoxy_products = (
-            [
-                (
-                    _ReactionComponent(p.name)
-                    if isinstance(p, Species)
-                    else _ReactionComponent(p[1].name, p[0])
-                )
-                for p in alkoxy_products
-            ]
-            if alkoxy_products is not None
-            else self.alkoxy_products
-        )
-        self.gas_phase = gas_phase.name if gas_phase is not None else self.gas_phase
-        self.other_properties = other_properties if other_properties is not None else self.other_properties
-
-    @staticmethod
-    def serialize(cls):
-        serialize_dict = {
-            "type": "BRANCHED_NO_RO2",
-            "name": cls.name,
-            "X": cls.X,
-            "Y": cls.Y,
-            "a0": cls.a0,
-            "n": cls.n,
-            "reactants": ReactionComponentSerializer.serialize_list_reaction_components(cls.reactants),
-            "nitrate products": ReactionComponentSerializer.serialize_list_reaction_components(cls.nitrate_products),
-            "alkoxy products": ReactionComponentSerializer.serialize_list_reaction_components(cls.alkoxy_products),
-            "gas phase": cls.gas_phase,
-        }
-        add_other_properties(serialize_dict, cls.other_properties)
-        return remove_empty_keys(serialize_dict)
-
-
-class Tunneling(_Tunneling):
-    """
-    A class representing a quantum tunneling reaction rate constant.
-
-    k = A * exp( -B / T ) * exp( C / T^3 )
-
-    where:
-        k = rate constant
-        A = pre-exponential factor [(mol m-3)^(n-1)s-1]
-        B = tunneling parameter [K^-1]
-        C = tunneling parameter [K^-3]
-        T = temperature [K]
-        n = number of reactants
-
-    Attributes:
-        name (str): The name of the tunneling reaction rate constant.
-        A (float): Pre-exponential factor [(mol m-3)^(n-1)s-1].
-        B (float): Tunneling parameter [K^-1].
-        C (float): Tunneling parameter [K^-3].
-        reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-        products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-        gas_phase (Phase): The gas phase in which the reaction occurs.
-        other_properties (Dict[str, Any]): A dictionary of other properties of the tunneling reaction rate constant.
-    """
-
-    def __init__(
-        self,
-        name: Optional[str] = None,
-        A: Optional[float] = None,
-        B: Optional[float] = None,
-        C: Optional[float] = None,
-        reactants: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        products: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        gas_phase: Optional[Phase] = None,
-        other_properties: Optional[Dict[str, Any]] = None,
-    ):
-        """
-        Initializes the Tunneling object with the given parameters.
-
-        Args:
-            name (str): The name of the tunneling reaction rate constant.
-            A (float): Pre-exponential factor [(mol m-3)^(n-1)s-1].
-            B (float): Tunneling parameter [K^-1].
-            C (float): Tunneling parameter [K^-3].
-            reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-            products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-            gas_phase (Phase): The gas phase in which the reaction occurs.
-            other_properties (Dict[str, Any]): A dictionary of other properties of the tunneling reaction rate constant.
-        """
-        super().__init__()
-        self.name = name if name is not None else self.name
-        self.A = A if A is not None else self.A
-        self.B = B if B is not None else self.B
-        self.C = C if C is not None else self.C
-        self.reactants = (
-            [
-                (
-                    _ReactionComponent(r.name)
-                    if isinstance(r, Species)
-                    else _ReactionComponent(r[1].name, r[0])
-                )
-                for r in reactants
-            ]
-            if reactants is not None
-            else self.reactants
-        )
-        self.products = (
-            [
-                (
-                    _ReactionComponent(p.name)
-                    if isinstance(p, Species)
-                    else _ReactionComponent(p[1].name, p[0])
-                )
-                for p in products
-            ]
-            if products is not None
-            else self.products
-        )
-        self.gas_phase = gas_phase.name if gas_phase is not None else self.gas_phase
-        self.other_properties = other_properties if other_properties is not None else self.other_properties
-
-    @staticmethod
-    def serialize(cls):
-        serialize_dict = {
-            "type": "TUNNELING",
-            "name": cls.name,
-            "A": cls.A,
-            "B": cls.B,
-            "C": cls.C,
-            "reactants": ReactionComponentSerializer.serialize_list_reaction_components(cls.reactants),
-            "products": ReactionComponentSerializer.serialize_list_reaction_components(cls.products),
-            "gas phase": cls.gas_phase,
-        }
-        add_other_properties(serialize_dict, cls.other_properties)
-        return remove_empty_keys(serialize_dict)
-
-
-class Surface(_Surface):
-    """
-    A class representing a surface in a chemical mechanism.
-
-    (TODO: get details from MusicBox)
-
-    Attributes:
-        name (str): The name of the surface.
-        reaction_probability (float): The probability of a reaction occurring on the surface.
-        gas_phase_species (Union[Species, Tuple[float, Species]]): The gas phase species involved in the reaction.
-        gas_phase_products (List[Union[Species, Tuple[float, Species]]]): The gas phase products formed in the reaction.
-        gas_phase (Phase): The gas phase in which the reaction occurs.
-        aerosol_phase (Phase): The aerosol phase in which the reaction occurs.
-        other_properties (Dict[str, Any]): A dictionary of other properties of the surface.
-    """
-
-    def __init__(
-        self,
-        name: Optional[str] = None,
-        reaction_probability: Optional[float] = None,
-        gas_phase_species: Optional[Union[Species, Tuple[float, Species]]] = None,
-        gas_phase_products: Optional[
-            List[Union[Species, Tuple[float, Species]]]
-        ] = None,
-        gas_phase: Optional[Phase] = None,
-        aerosol_phase: Optional[Phase] = None,
-        other_properties: Optional[Dict[str, Any]] = None,
-    ):
-        """
-        Initializes the Surface object with the given parameters.
-
-        Args:
-            name (str): The name of the surface.
-            reaction_probability (float): The probability of a reaction occurring on the surface.
-            gas_phase_species (Union[Species, Tuple[float, Species]]): The gas phase species involved in the reaction.
-            gas_phase_products (List[Union[Species, Tuple[float, Species]]]): The gas phase products formed in the reaction.
-            gas_phase (Phase): The gas phase in which the reaction occurs.
-            aerosol_phase (Phase): The aerosol phase in which the reaction occurs.
-            other_properties (Dict[str, Any]): A dictionary of other properties of the surface.
-        """
-        super().__init__()
-        self.name = name if name is not None else self.name
-        self.reaction_probability = reaction_probability if reaction_probability is not None else self.reaction_probability
-        self.gas_phase_species = (
-            (
-                _ReactionComponent(gas_phase_species.name)
-                if isinstance(gas_phase_species, Species)
-                else _ReactionComponent(gas_phase_species[1].name, gas_phase_species[0])
-            )
-            if gas_phase_species is not None
-            else self.gas_phase_species
-        )
-        self.gas_phase_products = (
-            [
-                (
-                    _ReactionComponent(p.name)
-                    if isinstance(p, Species)
-                    else _ReactionComponent(p[1].name, p[0])
-                )
-                for p in gas_phase_products
-            ]
-            if gas_phase_products is not None
-            else self.gas_phase_products
-        )
-        self.gas_phase = gas_phase.name if gas_phase is not None else self.gas_phase
-        self.aerosol_phase = aerosol_phase.name if aerosol_phase is not None else self.aerosol_phase
-        self.other_properties = other_properties if other_properties is not None else self.other_properties
-
-    @staticmethod
-    def serialize(cls):
-        serialize_dict = {
-            "type": "SURFACE",
-            "name": cls.name,
-            "reaction probability": cls.reaction_probability,
-            "gas-phase species": cls.gas_phase_species.species_name,
-            "gas-phase products": ReactionComponentSerializer.serialize_list_reaction_components(cls.gas_phase_products),
-            "gas phase": cls.gas_phase,
-            "aerosol phase": cls.aerosol_phase,
-        }
-        add_other_properties(serialize_dict, cls.other_properties)
-        return remove_empty_keys(serialize_dict)
-
-
-class Photolysis(_Photolysis):
-    """
-    A class representing a photolysis reaction rate constant.
-
-    Attributes:
-        name (str): The name of the photolysis reaction rate constant.
-        scaling_factor (float): The scaling factor for the photolysis rate constant.
-        reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-        products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-        gas_phase (Phase): The gas phase in which the reaction occurs.
-        other_properties (Dict[str, Any]): A dictionary of other properties of the photolysis reaction rate constant.
-    """
-
-    def __init__(
-        self,
-        name: Optional[str] = None,
-        scaling_factor: Optional[float] = None,
-        reactants: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        products: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        gas_phase: Optional[Phase] = None,
-        other_properties: Optional[Dict[str, Any]] = None,
-    ):
-        """
-        Initializes the Photolysis object with the given parameters.
-
-        Args:
-            name (str): The name of the photolysis reaction rate constant.
-            scaling_factor (float): The scaling factor for the photolysis rate constant.
-            reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-            products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-            gas_phase (Phase): The gas phase in which the reaction occurs.
-            other_properties (Dict[str, Any]): A dictionary of other properties of the photolysis reaction rate constant.
-        """
-        super().__init__()
-        self.name = name = name if name is not None else self.name
-        self.scaling_factor = scaling_factor if scaling_factor is not None else self.scaling_factor
-        self.reactants = (
-            [
-                (
-                    _ReactionComponent(r.name)
-                    if isinstance(r, Species)
-                    else _ReactionComponent(r[1].name, r[0])
-                )
-                for r in reactants
-            ]
-            if reactants is not None
-            else self.reactants
-        )
-        self.products = (
-            [
-                (
-                    _ReactionComponent(p.name)
-                    if isinstance(p, Species)
-                    else _ReactionComponent(p[1].name, p[0])
-                )
-                for p in products
-            ]
-            if products is not None
-            else self.products
-        )
-        self.gas_phase = gas_phase.name if gas_phase is not None else self.gas_phase
-        self.other_properties = other_properties if other_properties is not None else self.other_properties
-
-    @staticmethod
-    def serialize(cls):
-        serialize_dict = {
-            "type": "PHOTOLYSIS",
-            "name": cls.name,
-            "scaling factor": cls.scaling_factor,
-            "reactants": ReactionComponentSerializer.serialize_list_reaction_components(cls.reactants),
-            "products": ReactionComponentSerializer.serialize_list_reaction_components(cls.products),
-            "gas phase": cls.gas_phase,
-        }
-        add_other_properties(serialize_dict, cls.other_properties)
-        return remove_empty_keys(serialize_dict)
-
-
-class CondensedPhasePhotolysis(_CondensedPhasePhotolysis):
-    """
-    A class representing a condensed phase photolysis reaction rate constant.
-
-    Attributes:
-        name (str): The name of the condensed phase photolysis reaction rate constant.
-        scaling_factor (float): The scaling factor for the photolysis rate constant.
-        reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-        products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-        aerosol_phase (Phase): The aerosol phase in which the reaction occurs.
-        aerosol_phase_water (float): The water species in the aerosol phase [unitless].
-        other_properties (Dict[str, Any]): A dictionary of other properties of the condensed phase photolysis reaction rate constant.
-    """
-
-    def __init__(
-        self,
-        name: Optional[str] = None,
-        scaling_factor: Optional[float] = None,
-        reactants: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        products: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        aerosol_phase: Optional[Phase] = None,
-        aerosol_phase_water: Optional[Species] = None,
-        other_properties: Optional[Dict[str, Any]] = None,
-    ):
-        """
-        Initializes the CondensedPhasePhotolysis object with the given parameters.
-
-        Args:
-            name (str): The name of the condensed phase photolysis reaction rate constant.
-            scaling_factor (float): The scaling factor for the photolysis rate constant.
-            reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-            products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-            aerosol_phase (Phase): The aerosol phase in which the reaction occurs.
-            aerosol_phase_water (Species): The water species in the aerosol phase [unitless].
-            other_properties (Dict[str, Any]): A dictionary of other properties of the condensed phase photolysis reaction rate constant.
-        """
-        super().__init__()
-        self.name = name if name is not None else self.name
-        self.scaling_factor = scaling_factor if scaling_factor is not None else self.scaling_factor
-        self.reactants = (
-            [
-                (
-                    _ReactionComponent(r.name)
-                    if isinstance(r, Species)
-                    else _ReactionComponent(r[1].name, r[0])
-                )
-                for r in reactants
-            ]
-            if reactants is not None
-            else self.reactants
-        )
-        self.products = (
-            [
-                (
-                    _ReactionComponent(p.name)
-                    if isinstance(p, Species)
-                    else _ReactionComponent(p[1].name, p[0])
-                )
-                for p in products
-            ]
-            if products is not None
-            else self.products
-        )
-        self.aerosol_phase = aerosol_phase.name if aerosol_phase is not None else self.aerosol_phase
-        self.aerosol_phase_water = (
-            aerosol_phase_water.name if aerosol_phase_water is not None else self.aerosol_phase_water
-        )
-        self.other_properties = other_properties if other_properties is not None else self.other_properties
-
-    @staticmethod
-    def serialize(cls):
-        serialize_dict = {
-            "type": "CONDENSED_PHASE_PHOTOLYSIS",
-            "name": cls.name,
-            "scaling factor": cls.scaling_factor,
-            "reactants": ReactionComponentSerializer.serialize_list_reaction_components(cls.reactants),
-            "products": ReactionComponentSerializer.serialize_list_reaction_components(cls.products),
-            "aerosol phase": cls.aerosol_phase,
-            "aerosol-phase water": cls.aerosol_phase_water,
-        }
-        add_other_properties(serialize_dict, cls.other_properties)
-        return remove_empty_keys(serialize_dict)
-
-
-class Emission(_Emission):
-    """
-    A class representing an emission reaction rate constant.
-
-    Attributes:
-        name (str): The name of the emission reaction rate constant.
-        scaling_factor (float): The scaling factor for the emission rate constant.
-        products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-        gas_phase (Phase): The gas phase in which the reaction occurs.
-        other_properties (Dict[str, Any]): A dictionary of other properties of the emission reaction rate constant.
-    """
-
-    def __init__(
-        self,
-        name: Optional[str] = None,
-        scaling_factor: Optional[float] = None,
-        products: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        gas_phase: Optional[Phase] = None,
-        other_properties: Optional[Dict[str, Any]] = None,
-    ):
-        """
-        Initializes the Emission object with the given parameters.
-
-        Args:
-            name (str): The name of the emission reaction rate constant.
-            scaling_factor (float): The scaling factor for the emission rate constant.
-            products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-            gas_phase (Phase): The gas phase in which the reaction occurs.
-            other_properties (Dict[str, Any]): A dictionary of other properties of the emission reaction rate constant.
-        """
-        super().__init__()
-        self.name = name if name is not None else self.name
-        self.scaling_factor = scaling_factor if scaling_factor is not None else self.scaling_factor
-        self.products = (
-            [
-                (
-                    _ReactionComponent(p.name)
-                    if isinstance(p, Species)
-                    else _ReactionComponent(p[1].name, p[0])
-                )
-                for p in products
-            ]
-            if products is not None
-            else self.products
-        )
-        self.gas_phase = gas_phase.name if gas_phase is not None else self.gas_phase
-        self.other_properties = other_properties if other_properties is not None else self.other_properties
-
-    @staticmethod
-    def serialize(cls):
-        serialize_dict = {
-            "type": "EMISSION",
-            "name": cls.name,
-            "scaling factor": cls.scaling_factor,
-            "products": ReactionComponentSerializer.serialize_list_reaction_components(cls.products),
-            "gas phase": cls.gas_phase,
-        }
-        add_other_properties(serialize_dict, cls.other_properties)
-        return remove_empty_keys(serialize_dict)
-
-
-class FirstOrderLoss(_FirstOrderLoss):
-    """
-    A class representing a first-order loss reaction rate constant.
-
-    Attributes:
-        name (str): The name of the first-order loss reaction rate constant.
-        scaling_factor (float): The scaling factor for the first-order loss rate constant.
-        reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-        gas_phase (Phase): The gas phase in which the reaction occurs.
-        other_properties (Dict[str, Any]): A dictionary of other properties of the first-order loss reaction rate constant.
-    """
-
-    def __init__(
-        self,
-        name: Optional[str] = None,
-        scaling_factor: Optional[float] = None,
-        reactants: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        gas_phase: Optional[Phase] = None,
-        other_properties: Optional[Dict[str, Any]] = None,
-    ):
-        """
-        Initializes the FirstOrderLoss object with the given parameters.
-
-        Args:
-            name (str): The name of the first-order loss reaction rate constant.
-            scaling_factor (float): The scaling factor for the first-order loss rate constant.
-            reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-            gas_phase (Phase): The gas phase in which the reaction occurs.
-            other_properties (Dict[str, Any]): A dictionary of other properties of the first-order loss reaction rate constant.
-        """
-        super().__init__()
-        self.name = name if name is not None else self.name
-        self.scaling_factor = scaling_factor if scaling_factor is not None else self.scaling_factor
-        self.reactants = (
-            [
-                (
-                    _ReactionComponent(r.name)
-                    if isinstance(r, Species)
-                    else _ReactionComponent(r[1].name, r[0])
-                )
-                for r in reactants
-            ]
-            if reactants is not None
-            else self.reactants
-        )
-        self.gas_phase = gas_phase.name if gas_phase is not None else self.gas_phase
-        self.other_properties = other_properties if other_properties is not None else self.other_properties
-
-    @staticmethod
-    def serialize(cls):
-        serialize_dict = {
-            "type": "FIRST_ORDER_LOSS",
-            "name": cls.name,
-            "scaling factor": cls.scaling_factor,
-            "reactants": ReactionComponentSerializer.serialize_list_reaction_components(cls.reactants),
-            "gas phase": cls.gas_phase,
-        }
-        add_other_properties(serialize_dict, cls.other_properties)
-        return remove_empty_keys(serialize_dict)
-
-
-class AqueousEquilibrium(_AqueousEquilibrium):
-    """
-    A class representing an aqueous equilibrium reaction rate constant.
-
-    Attributes:
-        name (str): The name of the aqueous equilibrium reaction rate constant.
-        gas_phase (Phase): The gas phase in which the reaction occurs.
-        aerosol_phase (Phase): The aerosol phase in which the reaction occurs.
-        aerosol_phase_water (Species): The water species in the aerosol phase.
-        reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-        products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-        A (float): Pre-exponential factor [(mol m-3)^(n-1)s-1].
-        C (float): Exponential term [K-1].
-        k_reverse (float): Reverse rate constant [(mol m-3)^(n-1)s-1].
-        other_properties (Dict[str, Any]): A dictionary of other properties of the aqueous equilibrium reaction rate constant.
-    """
-
-    def __init__(
-        self,
-        name: Optional[str] = None,
-        aerosol_phase: Optional[Phase] = None,
-        aerosol_phase_water: Optional[Species] = None,
-        reactants: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        products: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        A: Optional[float] = None,
-        C: Optional[float] = None,
-        k_reverse: Optional[float] = None,
-        other_properties: Optional[Dict[str, Any]] = None,
-    ):
-        """
-        Initializes the AqueousEquilibrium object with the given parameters.
-
-        Args:
-            name (str): The name of the aqueous equilibrium reaction rate constant.
-            aerosol_phase (Phase): The aerosol phase in which the reaction occurs.
-            aerosol_phase_water (Species): The water species in the aerosol phase.
-            reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-            products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-            A (float): Pre-exponential factor [(mol m-3)^(n-1)s-1].
-            C (float): Exponential term [K-1].
-            k_reverse (float): Reverse rate constant [(mol m-3)^(n-1)s-1].
-            other_properties (Dict[str, Any]): A dictionary of other properties of the aqueous equilibrium reaction rate constant.
-        """
-        super().__init__()
-        self.name = name if name is not None else self.name
-        self.aerosol_phase = aerosol_phase.name if aerosol_phase is not None else self.aerosol_phase
-        self.aerosol_phase_water = (
-            aerosol_phase_water.name if aerosol_phase_water is not None else self.aerosol_phase_water
-        )
-        self.reactants = (
-            [
-                (
-                    _ReactionComponent(r.name)
-                    if isinstance(r, Species)
-                    else _ReactionComponent(r[1].name, r[0])
-                )
-                for r in reactants
-            ]
-            if reactants is not None
-            else self.reactants
-        )
-        self.products = (
-            [
-                (
-                    _ReactionComponent(p.name)
-                    if isinstance(p, Species)
-                    else _ReactionComponent(p[1].name, p[0])
-                )
-                for p in products
-            ]
-            if products is not None
-            else self.products
-        )
-        self.A = A if A is not None else self.A
-        self.C = C if C is not None else self.C
-        self.k_reverse = k_reverse if k_reverse is not None else self.k_reverse
-        self.other_properties = other_properties if other_properties is not None else self.other_properties
-
-    @staticmethod
-    def serialize(cls):
-        serialize_dict = {
-            "type": "AQUEOUS_EQUILIBRIUM",
-            "name": cls.name,
-            "aerosol phase": cls.aerosol_phase,
-            "aerosol-phase water": cls.aerosol_phase_water,
-            "reactants": ReactionComponentSerializer.serialize_list_reaction_components(cls.reactants),
-            "products": ReactionComponentSerializer.serialize_list_reaction_components(cls.products),
-            "A": cls.A,
-            "C": cls.C,
-            "k_reverse": cls.k_reverse,
-        }
-        add_other_properties(serialize_dict, cls.other_properties)
-        return remove_empty_keys(serialize_dict)
-
-
-class WetDeposition(_WetDeposition):
-    """
-    A class representing a wet deposition reaction rate constant.
-
-    Attributes:
-        name (str): The name of the wet deposition reaction rate constant.
-        scaling_factor (float): The scaling factor for the wet deposition rate constant.
-        aerosol_phase (Phase): The aerosol phase which undergoes wet deposition.
-        unknown_properties (Dict[str, Any]): A dictionary of other properties of the wet deposition reaction rate constant.
-    """
-
-    def __init__(
-        self,
-        name: Optional[str] = None,
-        scaling_factor: Optional[float] = None,
-        aerosol_phase: Optional[Phase] = None,
-        other_properties: Optional[Dict[str, Any]] = None,
-    ):
-        """
-        Initializes the WetDeposition object with the given parameters.
-
-        Args:
-            name (str): The name of the wet deposition reaction rate constant.
-            scaling_factor (float): The scaling factor for the wet deposition rate constant.
-            aerosol_phase (Phase): The aerosol phase which undergoes wet deposition.
-            other_properties (Dict[str, Any]): A dictionary of other properties of the wet deposition reaction rate constant.
-        """
-        super().__init__()
-        self.name = name if name is not None else self.name
-        self.scaling_factor = scaling_factor if scaling_factor is not None else self.scaling_factor
-        self.aerosol_phase = aerosol_phase.name if aerosol_phase is not None else self.aerosol_phase
-        self.other_properties = other_properties if other_properties is not None else self.other_properties
-
-    @staticmethod
-    def serialize(cls):
-        serialize_dict = {
-            "type": "WET_DEPOSITION",
-            "name": cls.name,
-            "scaling factor": cls.scaling_factor,
-            "aerosol phase": cls.aerosol_phase,
-        }
-        add_other_properties(serialize_dict, cls.other_properties)
-        return remove_empty_keys(serialize_dict)
-
-
-class HenrysLaw(_HenrysLaw):
-    """
-    A class representing a Henry's law reaction rate constant.
-
-    Attributes:
-        name (str): The name of the Henry's law reaction rate constant.
-        gas_phase (Phase): The gas phase in which the reaction occurs.
-        gas_phase_species (Union[Species, Tuple[float, Species]]): The gas phase species involved in the reaction.
-        aerosol_phase (Phase): The aerosol phase in which the reaction occurs.
-        aerosol_phase_water (Species): The water species in the aerosol phase.
-        aerosol_phase_species (Union[Species, Tuple[float, Species]]): The aerosol phase species involved in the reaction.
-        other_properties (Dict[str, Any]): A dictionary of other properties of the Henry's law reaction rate constant.
-    """
-
-    def __init__(
-        self,
-        name: Optional[str] = None,
-        gas_phase: Optional[Phase] = None,
-        gas_phase_species: Optional[Union[Species, Tuple[float, Species]]] = None,
-        aerosol_phase: Optional[Phase] = None,
-        aerosol_phase_water: Optional[Species] = None,
-        aerosol_phase_species: Optional[Union[Species, Tuple[float, Species]]] = None,
-        other_properties: Optional[Dict[str, Any]] = None,
-    ):
-        """
-        Initializes the HenrysLaw object with the given parameters.
-
-        Args:
-            name (str): The name of the Henry's law reaction rate constant.
-            gas_phase (Phase): The gas phase in which the reaction occurs.
-            gas_phase_species (Union[Species, Tuple[float, Species]]): The gas phase species involved in the reaction.
-            aerosol_phase (Phase): The aerosol phase in which the reaction occurs.
-            aerosol_phase_water (Species): The water species in the aerosol phase.
-            aerosol_phase_species (Union[Species, Tuple[float, Species]]): The aerosol phase species involved in the reaction.
-            other_properties (Dict[str, Any]): A dictionary of other properties of the Henry's law reaction rate constant.
-        """
-        super().__init__()
-        self.name = name if name is not None else self.name
-        self.gas_phase = gas_phase.name if gas_phase is not None else self.gas_phase
-        self.gas_phase_species = (
-            (
-                _ReactionComponent(gas_phase_species.name)
-                if isinstance(gas_phase_species, Species)
-                else _ReactionComponent(gas_phase_species[1].name, gas_phase_species[0])
-            )
-            if gas_phase_species is not None
-            else self.gas_phase_species
-        )
-        self.aerosol_phase = aerosol_phase.name if aerosol_phase is not None else self.aerosol_phase
-        self.aerosol_phase_water = (
-            aerosol_phase_water.name if aerosol_phase_water is not None else self.aerosol_phase_water
-        )
-        self.aerosol_phase_species = (
-            (
-                _ReactionComponent(aerosol_phase_species.name)
-                if isinstance(aerosol_phase_species, Species)
-                else _ReactionComponent(
-                    aerosol_phase_species[1].name, aerosol_phase_species[0]
-                )
-            )
-            if aerosol_phase_species is not None
-            else self.aerosol_phase_species
-        )
-        self.other_properties = other_properties if other_properties is not None else self.other_properties
-
-    @staticmethod
-    def serialize(cls):
-        serialize_dict = {
-            "type": "HL_PHASE_TRANSFER",
-            "name": cls.name,
-            "gas phase": cls.gas_phase,
-            "gas-phase species": cls.gas_phase_species.species_name,
-            "aerosol phase": cls.aerosol_phase,
-            "aerosol-phase water": cls.aerosol_phase_water,
-            "aerosol-phase species": cls.aerosol_phase_species.species_name,
-        }
-        add_other_properties(serialize_dict, cls.other_properties)
-        return remove_empty_keys(serialize_dict)
-
-
-class SimpolPhaseTransfer(_SimpolPhaseTransfer):
-    """
-    A class representing a simplified phase transfer reaction rate constant.
-
-    Attributes:
-        name (str): The name of the simplified phase transfer reaction rate constant.
-        gas_phase (Phase): The gas phase in which the reaction occurs.
-        gas_phase_species (Union[Species, Tuple[float, Species]]): The gas phase species involved in the reaction.
-        aerosol_phase (Phase): The aerosol phase in which the reaction occurs.
-        aerosol_phase_species (Union[Species, Tuple[float, Species]]): The aerosol phase species involved in the reaction.
-        B (List[float]): The B parameters [unitless].
-        unknown_properties (Dict[str, Any]): A dictionary of other properties of the simplified phase transfer reaction rate constant.
-    """
-
-    def __init__(
-        self,
-        name: Optional[str] = None,
-        gas_phase: Optional[Phase] = None,
-        gas_phase_species: Optional[Union[Species, Tuple[float, Species]]] = None,
-        aerosol_phase: Optional[Phase] = None,
-        aerosol_phase_species: Optional[Union[Species, Tuple[float, Species]]] = None,
-        B: Optional[List[float]] = None,
-        other_properties: Optional[Dict[str, Any]] = None,
-    ):
-        """
-        Initializes the SimpolPhaseTransfer object with the given parameters.
-
-        Args:
-            name (str): The name of the simplified phase transfer reaction rate constant.
-            gas_phase (Phase): The gas phase in which the reaction occurs.
-            gas_phase_species (Union[Species, Tuple[float, Species]]): The gas phase species involved in the reaction.
-            aerosol_phase (Phase): The aerosol phase in which the reaction occurs.
-            aerosol_phase_species (Union[Species, Tuple[float, Species]]): The aerosol phase species involved in the reaction.
-            B (List[float]): The B parameters [unitless].
-            other_properties (Dict[str, Any]): A dictionary of other properties of the simplified phase transfer reaction rate constant.
-        """
-        super().__init__()
-        self.name = name if name is not None else self.name
-        self.gas_phase = gas_phase.name if gas_phase is not None else self.gas_phase
-        self.gas_phase_species = (
-            (
-                _ReactionComponent(gas_phase_species.name)
-                if isinstance(gas_phase_species, Species)
-                else _ReactionComponent(gas_phase_species[1].name, gas_phase_species[0])
-            )
-            if gas_phase_species is not None
-            else self.gas_phase_species
-        )
-        self.aerosol_phase = aerosol_phase.name if aerosol_phase is not None else self.aerosol_phase
-        self.aerosol_phase_species = (
-            (
-                _ReactionComponent(aerosol_phase_species.name)
-                if isinstance(aerosol_phase_species, Species)
-                else _ReactionComponent(
-                    aerosol_phase_species[1].name, aerosol_phase_species[0]
-                )
-            )
-            if aerosol_phase_species is not None
-            else self.aerosol_phase_species
-        )
-        if B is not None:
-            if len(B) != 4:
-                raise ValueError("B must be a list of 4 elements.")
-            self.B = B
-        else:
-            self.B = [0, 0, 0, 0]
-        self.other_properties = other_properties if other_properties is not None else self.other_properties
-
-    @staticmethod
-    def serialize(cls):
-        serialize_dict = {
-            "type": "SIMPOL_PHASE_TRANSFER",
-            "name": cls.name,
-            "gas phase": cls.gas_phase,
-            "gas-phase species": cls.gas_phase_species.species_name,
-            "aerosol phase": cls.aerosol_phase,
-            "aerosol-phase species": cls.aerosol_phase_species.species_name,
-            "B": cls.B,
-        }
-        add_other_properties(serialize_dict, cls.other_properties)
-        return remove_empty_keys(serialize_dict)
-
-
-class UserDefined(_UserDefined):
-    """
-    A class representing a user-defined reaction rate constant.
-
-    Attributes:
-        name (str): The name of the photolysis reaction rate constant.
-        scaling_factor (float): The scaling factor for the photolysis rate constant.
-        reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-        products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-        gas_phase (Phase): The gas phase in which the reaction occurs.
-        other_properties (Dict[str, Any]): A dictionary of other properties of the photolysis reaction rate constant.
-    """
-
-    def __init__(
-        self,
-        name: Optional[str] = None,
-        scaling_factor: Optional[float] = None,
-        reactants: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        products: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
-        gas_phase: Optional[Phase] = None,
-        other_properties: Optional[Dict[str, Any]] = None,
-    ):
-        """
-        Initializes the UserDefined object with the given parameters.
-
-        Args:
-            name (str): The name of the photolysis reaction rate constant.
-            scaling_factor (float): The scaling factor for the photolysis rate constant.
-            reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
-            products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
-            gas_phase (Phase): The gas phase in which the reaction occurs.
-            other_properties (Dict[str, Any]): A dictionary of other properties of the photolysis reaction rate constant.
-        """
-        super().__init__()
-        self.name = name if name is not None else self.name
-        self.scaling_factor = scaling_factor if scaling_factor is not None else self.scaling_factor
-        self.reactants = (
-            [
-                (
-                    _ReactionComponent(r.name)
-                    if isinstance(r, Species)
-                    else _ReactionComponent(r[1].name, r[0])
-                )
-                for r in reactants
-            ]
-            if reactants is not None
-            else self.reactants
-        )
-        self.products = (
-            [
-                (
-                    _ReactionComponent(p.name)
-                    if isinstance(p, Species)
-                    else _ReactionComponent(p[1].name, p[0])
-                )
-                for p in products
-            ]
-            if products is not None
-            else self.products
-        )
-        self.gas_phase = gas_phase.name if gas_phase is not None else self.gas_phase
-        self.other_properties = other_properties if other_properties is not None else self.other_properties
-
-    @staticmethod
-    def serialize(cls):
-        serialize_dict = {
-            "type": "USER_DEFINED",
-            "name": cls.name,
-            "scaling factor": cls.scaling_factor,
-            "reactants": ReactionComponentSerializer.serialize_list_reaction_components(cls.reactants),
-            "products": ReactionComponentSerializer.serialize_list_reaction_components(cls.products),
-            "gas phase": cls.gas_phase,
-        }
-        add_other_properties(serialize_dict, cls.other_properties)
-        return remove_empty_keys(serialize_dict)
+from musica.mechanism_configuration_arrhenius import Arrhenius, _Arrhenius
+from musica.mechanism_configuration_condensed_phase_arrhenius import CondensedPhaseArrhenius, _CondensedPhaseArrhenius
+from musica.mechanism_configuration_troe import Troe, _Troe
+from musica.mechanism_configuration_branched import Branched, _Branched
+from musica.mechanism_configuration_tunneling import Tunneling, _Tunneling
+from musica.mechanism_configuration_surface import Surface, _Surface
+from musica.mechanism_configuration_photolysis import Photolysis, _Photolysis
+from musica.mechanism_configuration_condensed_phase_photolysis import CondensedPhasePhotolysis, _CondensedPhasePhotolysis
+from musica.mechanism_configuration_emission import Emission, _Emission
+from musica.mechanism_configuration_first_order_loss import FirstOrderLoss, _FirstOrderLoss
+from musica.mechanism_configuration_aqueous_equilibrium import AqueousEquilibrium, _AqueousEquilibrium
+from musica.mechanism_configuration_wet_deposition import WetDeposition, _WetDeposition
+from musica.mechanism_configuration_henrys_law import HenrysLaw, _HenrysLaw
+from musica.mechanism_configuration_simpol_phase_transfer import SimpolPhaseTransfer, _SimpolPhaseTransfer
+from musica.mechanism_configuration_user_defined import UserDefined, _UserDefined
+
+from musica.mechanism_configuration_reactions import Reactions, ReactionType
 
 
 class Version(_Version):
@@ -1391,7 +97,7 @@ class Mechanism(_Mechanism):
         self.reactions = Reactions(reactions=reactions if reactions is not None else [])
         self.version = version if version is not None else Version()
 
-    def to_dict(self):
+    def to_dict(self) -> Dict:
         species_list = []
         for species in self.species:
             species_list.append(Species.serialize(species))
