+++ musica/mechanism_configuration_aqueous_equilibrium.py
@@ -1,0 +1,101 @@
+from typing import Optional, Any, Dict, List, Union, Tuple
+from musica import _AqueousEquilibrium, _ReactionComponent
+from musica.mechanism_configuration_phase import Phase
+from musica.mechanism_configuration_species import Species
+from musica.mechanism_configuration_reactions import ReactionComponentSerializer
+from musica.mechanism_configuration_utils import add_other_properties, remove_empty_keys
+
+
+class AqueousEquilibrium(_AqueousEquilibrium):
+    """
+    A class representing an aqueous equilibrium reaction rate constant.
+
+    Attributes:
+        name (str): The name of the aqueous equilibrium reaction rate constant.
+        gas_phase (Phase): The gas phase in which the reaction occurs.
+        aerosol_phase (Phase): The aerosol phase in which the reaction occurs.
+        aerosol_phase_water (Species): The water species in the aerosol phase.
+        reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
+        products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
+        A (float): Pre-exponential factor [(mol m-3)^(n-1)s-1].
+        C (float): Exponential term [K-1].
+        k_reverse (float): Reverse rate constant [(mol m-3)^(n-1)s-1].
+        other_properties (Dict[str, Any]): A dictionary of other properties of the aqueous equilibrium reaction rate constant.
+    """
+
+    def __init__(
+        self,
+        name: Optional[str] = None,
+        aerosol_phase: Optional[Phase] = None,
+        aerosol_phase_water: Optional[Species] = None,
+        reactants: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
+        products: Optional[List[Union[Species, Tuple[float, Species]]]] = None,
+        A: Optional[float] = None,
+        C: Optional[float] = None,
+        k_reverse: Optional[float] = None,
+        other_properties: Optional[Dict[str, Any]] = None,
+    ):
+        """
+        Initializes the AqueousEquilibrium object with the given parameters.
+
+        Args:
+            name (str): The name of the aqueous equilibrium reaction rate constant.
+            aerosol_phase (Phase): The aerosol phase in which the reaction occurs.
+            aerosol_phase_water (Species): The water species in the aerosol phase.
+            reactants (List[Union[Species, Tuple[float, Species]]]): A list of reactants involved in the reaction.
+            products (List[Union[Species, Tuple[float, Species]]]): A list of products formed in the reaction.
+            A (float): Pre-exponential factor [(mol m-3)^(n-1)s-1].
+            C (float): Exponential term [K-1].
+            k_reverse (float): Reverse rate constant [(mol m-3)^(n-1)s-1].
+            other_properties (Dict[str, Any]): A dictionary of other properties of the aqueous equilibrium reaction rate constant.
+        """
+        super().__init__()
+        self.name = name if name is not None else self.name
+        self.aerosol_phase = aerosol_phase.name if aerosol_phase is not None else self.aerosol_phase
+        self.aerosol_phase_water = (
+            aerosol_phase_water.name if aerosol_phase_water is not None else self.aerosol_phase_water
+        )
+        self.reactants = (
+            [
+                (
+                    _ReactionComponent(r.name)
+                    if isinstance(r, Species)
+                    else _ReactionComponent(r[1].name, r[0])
+                )
+                for r in reactants
+            ]
+            if reactants is not None
+            else self.reactants
+        )
+        self.products = (
+            [
+                (
+                    _ReactionComponent(p.name)
+                    if isinstance(p, Species)
+                    else _ReactionComponent(p[1].name, p[0])
+                )
+                for p in products
+            ]
+            if products is not None
+            else self.products
+        )
+        self.A = A if A is not None else self.A
+        self.C = C if C is not None else self.C
+        self.k_reverse = k_reverse if k_reverse is not None else self.k_reverse
+        self.other_properties = other_properties if other_properties is not None else self.other_properties
+
+    @staticmethod
+    def serialize(cls) -> Dict:
+        serialize_dict = {
+            "type": "AQUEOUS_EQUILIBRIUM",
+            "name": cls.name,
+            "aerosol phase": cls.aerosol_phase,
+            "aerosol-phase water": cls.aerosol_phase_water,
+            "reactants": ReactionComponentSerializer.serialize_list_reaction_components(cls.reactants),
+            "products": ReactionComponentSerializer.serialize_list_reaction_components(cls.products),
+            "A": cls.A,
+            "C": cls.C,
+            "k_reverse": cls.k_reverse,
+        }
+        add_other_properties(serialize_dict, cls.other_properties)
+        return remove_empty_keys(serialize_dict)
